REFACTOR: Moved messages and tests of qgb module to own file.
